Assert that the object contains exp list
Asserting that object["public"] equals the expected list fails when the
local database is not empty. e.g. if the user creates a table,
object["public"].Tables contains extraneous elements the expected tables
as well as the user-created table. In this case the equality assertion
fails. A correct test here would be asserting that the returned
table/seq/view list CONTAINS the exp list.

Also remove mapKeys() since it is no longer used anywhere else.